# FSIM Optimisation to Reduce CPU Usage