--- a/piq/functional/filters.py
+++ b/piq/functional/filters.py
@@ -1,9 +1,10 @@
 r"""Filters for gradient computation, bluring, etc."""
 import torch
 import numpy as np
+from typing import Optional
 
 
-def haar_filter(kernel_size: int, device: str = None, dtype: type = None) -> torch.Tensor:
+def haar_filter(kernel_size: int, device: Optional[str] = None, dtype: Optional[type] = None) -> torch.Tensor:
     r"""Creates Haar kernel
 
     Args:
@@ -18,7 +19,7 @@ def haar_filter(kernel_size: int, device: str = None, dtype: type = None) -> tor
     return kernel.unsqueeze(0)
 
 
-def hann_filter(kernel_size: int, device: str = None, dtype: type = None) -> torch.Tensor:
+def hann_filter(kernel_size: int, device: Optional[str] = None, dtype: Optional[type] = None) -> torch.Tensor:
     r"""Creates  Hann kernel
     Args:
         kernel_size: size of the kernel
@@ -34,7 +35,8 @@ def hann_filter(kernel_size: int, device: str = None, dtype: type = None) -> tor
     return kernel.view(1, kernel_size, kernel_size) / kernel.sum()
 
 
-def gaussian_filter(kernel_size: int, sigma: float, device: str = None, dtype: torch.dtype = None) -> torch.Tensor:
+def gaussian_filter(kernel_size: int, sigma: float, device: Optional[str] = None,
+                    dtype: Optional[type] = None) -> torch.Tensor:
     r"""Returns 2D Gaussian kernel N(0,`sigma`^2)
     Args:
         size: Size of the kernel
@@ -55,7 +57,7 @@ def gaussian_filter(kernel_size: int, sigma: float, device: str = None, dtype: t
 
 
 # Gradient operator kernels
-def scharr_filter(device: str = None, dtype: type = None) -> torch.Tensor:
+def scharr_filter(device: Optional[str] = None, dtype: Optional[type] = None) -> torch.Tensor:
     r"""Utility function that returns a normalized 3x3 Scharr kernel in X direction
 
     Args:
@@ -67,7 +69,7 @@ def scharr_filter(device: str = None, dtype: type = None) -> torch.Tensor:
     return torch.tensor([[[-3., 0., 3.], [-10., 0., 10.], [-3., 0., 3.]]], device=device, dtype=dtype) / 16
 
 
-def prewitt_filter(device: str = None, dtype: type = None) -> torch.Tensor:
+def prewitt_filter(device: Optional[str] = None, dtype: Optional[type] = None) -> torch.Tensor:
     r"""Utility function that returns a normalized 3x3 Prewitt kernel in X direction
 
     Args:
@@ -78,7 +80,7 @@ def prewitt_filter(device: str = None, dtype: type = None) -> torch.Tensor:
     return torch.tensor([[[-1., 0., 1.], [-1., 0., 1.], [-1., 0., 1.]]], device=device, dtype=dtype) / 3
 
 
-def binomial_filter1d(kernel_size: int, device: str = None, dtype: type = None) -> torch.Tensor:
+def binomial_filter1d(kernel_size: int, device: Optional[str] = None, dtype: Optional[type] = None) -> torch.Tensor:
     r"""Creates 1D normalized binomial filter
 
     Args:
@@ -93,7 +95,7 @@ def binomial_filter1d(kernel_size: int, device: str = None, dtype: type = None) 
     return torch.tensor(kernel.c, dtype=dtype, device=device).view(1, 1, kernel_size)
 
 
-def average_filter2d(kernel_size: int, device: str = None, dtype: type = None) -> torch.Tensor:
+def average_filter2d(kernel_size: int, device: Optional[str] = None, dtype: Optional[type] = None) -> torch.Tensor:
     r"""Creates 2D normalized average filter
 
     Args: