Closes #266 
Addresses the second part of #227 

## Proposed Changes
- FSIM: got rid of a duplicate for filter generation `_construct_filters`
- FSIM: removed the duplicate of mesh grid generation performed for `_lowpassfilter`. The filter is computed explicitly.
- FSIM: minor optimisation of tensor generations across the `fsim` to avoid transfer between devices
- base.py: mesh grid creation (`get_meshgrid`) is performed straight on target device and data type according to input tensor without redundant transfers between devices
- base.py: removed redundant `to()` operation in `gradient_map`
- filters: filters are generated straight on target device and data type according to input tensor without redundant transfers between devices
- colour_conversion: moved tensor generation to the target device instead of transfer between devices
- Updated `.gitignore`
- minor change in benchmark script to allow more flexible choice for GPU device
- minor fixes across the code to match the changes for colour conversion and filter generation

## Outcome of the Proposed Changes

- Reduced cpu usage with `cuda`-enabled computations for FSIM and FSIMc (visual comparison is atttached)
- CPU time was reduced by 6 sec according to PyTorch profiler

## Note
Other places to consider for additional assessment of leaks
- `piq/base.py`,  `fid.py`, `perceptual.py` and `pieapp.py` use `.to()` for a various transfers. They are out of scope for current PR due to a different purpose of the transfer between devices/types.
- custom made `piq/functional/resize.py` could be refactored as well to get rid of redundant transfers. Added to #227 
- `vsi.py` has redundant mesh grid creation. Opened #335 


# Appendix 
## Visual Comparison of CPU usage
`htop` following #266 

### Before
Branch: master
<img width="981" alt="Screenshot 2023-02-05 at 00 42 23" src="https://user-images.githubusercontent.com/22414094/216795789-d1442ad7-3093-4a65-bf58-3007a773da1f.png">

### After
Branch: bug/cpu_usage
<img width="1016" alt="Screenshot 2023-02-05 at 00 38 06" src="https://user-images.githubusercontent.com/22414094/216795795-e9153ed0-eefe-44c0-9374-f1a4aa780c92.png">



## PyTorch Profiler Results. Top-8 operations.
```
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  
                                                   Name    Self CPU %      Self CPU   CPU total %     CPU total  CPU time avg    # of Calls  
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  
                                        model_inference        16.38%        8.624s       100.00%       52.639s       52.639s             1  
                                               aten::to         0.07%      37.231ms        68.84%       36.234s       1.020ms         35525  
                                         aten::_to_copy         0.14%      72.795ms        68.76%       36.197s       4.993ms          7250  
                                            aten::copy_         0.34%     177.512ms        61.16%       32.196s       2.019ms         15950  
                                  cudaStreamSynchronize        40.22%       21.173s        40.22%       21.173s       1.718ms         12325  
                                        cudaMemcpyAsync        20.68%       10.883s        20.68%       10.883s     883.015us         12325  
                                    aten::empty_strided         0.17%      88.161ms         7.78%        4.093s     564.510us          7250  
                                             cudaMalloc         7.62%        4.009s         7.62%        4.009s     129.312ms            31  
```